перенес ветку в меню сообщения, доделать верстку векти, указать сообщение
