+++ source/entities/ChatMessage/ui/ChatMessage.handlebars
@@ -18,14 +18,6 @@
       <span class="timedate">
         {{message.datetime}}
       </span>
-      {{#if chatIsNotBranch}}
-      <div id="branch" class="message-branch">
-        <img class="message-branch__img" src="/assets/image/branch-icon.png" alt="">
-        <span class="message-branch__span">
-          100  
-        </span>
-      </div>
-      {{/if}}
     </span>
   </div>
