+++ source/widgets/MessageMenu/ui/MessageMenu.ts
@@ -1,6 +1,11 @@
+import { ChatStorage } from "@/entities/Chat/lib/ChatStore";
 import MessageMenuTemplate from "./MessageMenu.handlebars";
 import "./MessageMenu.scss";
 import { DeleteMessage } from "@/feature/DeleteMessage";
+import { ChatMessage, TChatMessage } from "@/entities/ChatMessage";
+import { API } from "@/shared/api/api";
+import { ChatMessagesResponse, createBranchResponse, EmptyRequest } from "@/shared/api/types";
+import { event } from "jquery";
 
 export class MessageMenu {
     #parent;
@@ -8,8 +13,10 @@ export class MessageMenu {
         this.#parent = parent;
     }
 
-    render(messageId : string, messageText : string, x : number, y : number, branch : boolean = false) {
-        this.#parent.innerHTML = MessageMenuTemplate({x, y});
+    render(message : TChatMessage, messageId : string, messageText : string, x : number, y : number, chatMessageObject : ChatMessage, branch : boolean = false) {
+      console.log(message)  
+      const notBranch = !branch;
+        this.#parent.innerHTML = MessageMenuTemplate({x, y, notBranch});
         const deleteButton = this.#parent.querySelector("#delete-message")!;
 
         const handleDelete = async () => {
@@ -50,5 +57,65 @@ export class MessageMenu {
           };
       
           document.addEventListener("click", handlerClickOutsideModal);
+
+    
+        const branchMessage = document.querySelector("#branch-message")!;
+        
+        const handleOpenBranch = async (event : MouseEvent) => {
+          event.stopPropagation();
+          this.#parent.innerHTML = '';
+            const startBranch = document.getElementById('start-branch')!;
+            const branchInput = document.getElementById("branch-input")!;
+            if (message?.branchId) {
+                ChatStorage.setCurrentBranchId(message.branchId);
+                startBranch?.classList?.add("hidden");
+                branchInput?.classList?.remove("hidden");
+            }
+            else {
+                startBranch?.classList?.remove("hidden");
+                branchInput?.classList?.add("hidden");
+            }
+            const currentChat = document.getElementById('chat')!;
+            const branchChat = document.getElementById("chat-branch")!;
+            currentChat?.classList.add("hidden");
+            branchChat?.classList.remove("hidden");
+            const chatBranch = document.getElementById("chat-branch")!;
+            const chatBranchMessages : HTMLElement = chatBranch.querySelector("#chat__messages")!;
+            chatBranchMessages.innerText = '';
+
+            const handleStartBranch = async () => {
+                const response = await API.post<createBranchResponse, EmptyRequest>(`/chat/${message.chatId}/${message.messageId}/branch`, {});
+                console.log(response);
+                if (!response.error) {
+                    ChatStorage.setCurrentBranchId(response.id);
+                    startBranch?.classList.add("hidden");
+                    branchInput?.classList.remove("hidden");
+                    const chatBranch = document.querySelector("#chat-branch")!;
+                    message.branchId = response.id;
+                    chatMessageObject.setParent(chatBranch.querySelector("#chat__messages")!);
+                    
+                }
+                return;
+            };
+
+            
+            if (message?.branchId) {
+                
+                ChatStorage.setCurrentBranchId(message.branchId);
+                const branchMessages = await API.get<ChatMessagesResponse>(`/chat/${message.branchId}/messages`);
+                
+                if (!branchMessages.error) {
+                  chatMessageObject.setParent(chatBranchMessages);
+                  chatMessageObject.renderMessages(branchMessages.messages, false);
+                }
+            }
+            if (startBranch) {
+                startBranch.addEventListener('click', handleStartBranch);
+            }
+        };
+        if (branchMessage) {
+          branchMessage.addEventListener("click", handleOpenBranch);
+        }
+        
     }
 }
